uhd: set both Rx and Tx freqs with use of tags and msgs
Current implementation of controlling frequency in both directions,
with use of single UHD USRP Source or Sink blocks and "direction"
key, doesn't allow to set Tx and Rx frequencies simultanously.

This change adds such ability, but it doesn't solve another issue which
is assumption that Tx and Rx block have the same number of channels.
Fixing that issue wasn't as easy so it was left to be taken care of
in the future.